fix: make color nullable and int, add tags to dashboard
