+++ database/studyu-schema.sql
@@ -133,7 +133,7 @@ ALTER TABLE public.study_subject OWNER TO supabase_admin;
 CREATE TABLE study_tag (
     id uuid DEFAULT gen_random_uuid() NOT NULL,
     name text NOT NULL,
-    color text NOT NULL,
+    color integer,
     parent_id uuid,
 );
